Helps the User Search For Tracks.
This Feature allows the users to search for tracks on the musix match
database using either/both track name, artist name, or
a part of the lyrics. It returns the results with track ID's which are useful when searching
for their favorite song's lyrics.
Data is returned in a tabular way for presentability.

ref #6