A functionality to allow users to access song's lyrics using track_id
This commit adds functionality to the code to ensure that a user can
get the lyrics for a specific song using the song's musixmatch track_id
more to that, it removed the necessity for importing authentification
to avoid creation of cache files that were affecting perfomance.
The user can copy and paste a track_id into a function and thus get the lyrics.

ref #8